refactor: move validator interface to sdk
