+++ packages/adapter-ccxt/src/validation/index.ts
@@ -1,4 +1,3 @@
-export * from "./IValidator";
 export * from "./AdapterConfigValidator";
 export * from "./StepConfigValidator";
 export * from "./ValidationError";
