+++ packages/sdk/src/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from "./IAdapter";
 export * from "./IConfigProvider";
+export * from "./IValidator";
